+++ crates/vertigo/src/dom/css.rs
@@ -5,6 +5,10 @@ pub enum CssGroup {
     CssStatic { value: &'static str },
     // string in this case, is a key to hashmap with the class name
     CssDynamic { value: String },
+    CssMedia {
+        query: String,
+        rules: Vec<String>,
+    }
 }
 
 /// CSS styles definition for Virtual DOM.