+++ crates/vertigo/src/html_macro/tests/css/animation.rs
@@ -1,13 +1,12 @@
-use crate::css_fn;
-
+use crate::css;
 use super::utils::*;
 
-// Make crate available by its name for css_fn macro
+// Make crate available by its name for css macro
 use crate as vertigo;
 
 #[test]
-fn animation_css() {
-    css_fn! { css_factory, "
+fn test_animation() {
+    let value = css!("
         animation: 5s linear 2s infinite alternate {
             0%   {background-color: red; left: 0px; top: 0px;}
             25%  {background-color: yellow; left: 200px; top: 0px;}
@@ -15,9 +14,7 @@ fn animation_css() {
             75%  {background-color: green; left: 0px; top: 200px;}
             100% {background-color: red; left: 0px; top: 0px;}
         };
-    " }
-
-    let value = css_factory();
+    ");
 
     assert_eq!(get_s(&value), "animation: 5s linear 2s infinite alternate { 0% { background-color: red;
 left: 0px;