+++ demo/app/src/app/github_explorer/component.rs
@@ -1,4 +1,4 @@
-use vertigo::{css_fn, Resource, bind, DomElement, dom, Computed, DomCommentCreate, transaction};
+use vertigo::{css, Resource, bind, DomElement, dom, Computed, DomCommentCreate, transaction};
 
 use super::State;
 
@@ -23,10 +23,25 @@ impl GitHubExplorer {
             });
         });
 
+        let wrapper = css!("
+            border: 1px solid black;
+            margin: 20px 0;
+            padding: 10px;
+        ");
+
+        let input_css = css!("
+            margin-left: 10px;
+        ");
+
+        let button_css = || css!("
+            margin: 0 10px;
+            cursor: pointer;
+        ");
+
         dom! {
-            <div css={wrapper()}>
+            <div css={wrapper}>
                 "Enter author/repo tuple: "
-                <input css={input_css()} value={state.repo_input.to_computed()} on_input={on_input_callback} />
+                <input css={input_css} value={state.repo_input.to_computed()} on_input={on_input_callback} />
                 <button css={button_css()} on_click={on_show}>"Fetch"</button>
                 <div css={button_css()}>
                     <text computed={&state.repo_shown} />
@@ -54,34 +69,19 @@ impl GitHubExplorer {
         });
 
         commit_message.render_value(|message| {
+            let text_css = css!("
+                width: 600px;
+                height: 300px;
+                border: 1px solid black;
+                padding: 5px;
+                margin: 10px;
+            ");
+
             dom! {
-                <div css={text_css()}>
+                <div css={text_css}>
                     { message }
                 </div>
             }
         })
     }
 }
-
-css_fn! { wrapper, "
-    border: 1px solid black;
-    margin: 20px 0;
-    padding: 10px;
-" }
-
-css_fn! { input_css, "
-    margin-left: 10px;
-" }
-
-css_fn! { button_css, "
-    margin: 0 10px;
-    cursor: pointer;
-" }
-
-css_fn! { text_css, "
-    width: 600px;
-    height: 300px;
-    border: 1px solid black;
-    padding: 5px;
-    margin: 10px;
-" }